requirements.txt
click>=8.0
PyYAML>=6.0
ollama-python>=0.1     # if there’s a wrapper lib; otherwise subprocess is enough